Ajustar ConfigurarActivity

+++ app/src/main/java/com/teconsis/monitora/Dispositivo.kt
@@ -3,6 +3,5 @@ package com.teconsis.monitora
 data class Dispositivo(
     val id: Int,
     val descricao: String,
-    val idAparelho: Long
+    val idAparelho: Int
 )
-
